move execute msg handling to struct
