+++ contracts/cw-ics721-bridge/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod contract;
 mod error;
+mod execute;
 pub mod ibc;
 pub mod ibc_helpers;
 pub mod ibc_packet_receive;
